Add 'mv' command & minor fixes
The command is achieved by first searching for the object in
the given source directory, and if found, removed and copied
to the dest directory. If the dest directory already has an object
with the same name, then replace it. There is also some extensive
error handling to catch all the possible errors, for example
moving a directory in itself, or moving a directory in a place
where there is a name collision.

Signed-off-by: Panagiotis Foliadis <pfoliadis@hotmail.com>