Add `cp` command
The way `cp` command was utilized is similar to the `mv` command,
but only for `FSFILE` types and not directories.

Also this commit contains a small refactor to simplify the creation
of an object, by moving to the constructor the responsibility
of setting the parent directory. Also a constructor was added
for the special case of the Root directory, since it doesn't
have a parent directory.

Signed-off-by: Panagiotis Foliadis <pfoliadis@hotmail.com>